Bug fix cache directives
